Indoor staging

+++ app/Http/Controllers/NewsletterTemplateController.php
@@ -13,22 +13,33 @@ class NewsletterTemplateController extends Controller
 {
     public function index()
     {
-        $templates = NewsletterTemplate::all();
+        $templates = NewsletterTemplate::with('sent_newsletter' , 'sent_newsletter.subscriber_email_list' , 'sent_newsletter.subscriber_email_list.subscriberEmailList')->orderBy('created_at' , 'DESC')->get();
         return view('newsletter_layout.newsletter_templates.index', compact('templates'));
     }
 
     public function create()
     {
-        $subscriber_email_lists = SubscriberList::all();
-        return view('newsletter_layout.newsletter_templates.create', compact('subscriber_email_lists'));
+        $subscriber_email_lists = SubscriberList::with('subscriberEmailList')->orderBy('id' , 'Desc')->get();
+        if (count($subscriber_email_lists) ==  0 ) {
+            return redirect()->route('subscribers_list_create')->with('error', 'Please create a subscriber list first!');
+        }
+        else {
+            return view('newsletter_layout.newsletter_templates.create', compact('subscriber_email_lists'));
+        }
     }
 
     public function store(Request $request)
     {
-        $request->validate([
-            'name' => 'required',
-            'content' => 'required',
-        ]);
+        $request->validate(
+            [
+                'name' => 'required',
+                'content' => 'required',
+                'subscriber_email_list_id' => 'required'
+            ] , 
+            [
+                'subscriber_email_list_id.required' => 'Please select a subscriber list!'
+            ]
+        );
         $newsletter_template = new NewsletterTemplate();
         $newsletter_template->name = $request->name;
         $newsletter_template->content = $request->content;
@@ -68,11 +79,19 @@ class NewsletterTemplateController extends Controller
         return view('newsletter_layout.newsletter_templates.edit', compact('newsletterTemplate' , 'subscriber_email_lists' , 'selected_id'));
     }
 
-    public function duplicate_newsletter_template($id) {
+    public function duplicate_newsletter_template(Request $request , $id) {
         $newsletterTemplate = NewsletterTemplate::find($id);
         $newNewsletterTemplate = $newsletterTemplate->replicate();
         $newNewsletterTemplate->name = $newsletterTemplate->name . ' (Copy)';
         $newNewsletterTemplate->save();
+
+
+        if (!empty($request->subscriber_list_id)) {
+            $newsletter_subscriber_template = new NewsletterSubscriberTemplate();
+            $newsletter_subscriber_template->list_id = $request->subscriber_list_id;
+            $newsletter_subscriber_template->newsletter_template_id = $newNewsletterTemplate->id;
+            $newsletter_subscriber_template->save();
+        }
 
         return redirect()->route('newsletter-templates.index')->with('success', 'Newsletter template duplicated successfully!');
     }
@@ -112,7 +131,9 @@ class NewsletterTemplateController extends Controller
     {
         $newsletterTemplate = NewsletterTemplate::find($id);
         $delete_newsletter_subscriber_template = NewsletterSubscriberTemplate::where('newsletter_template_id', $newsletterTemplate->id)->first();
-        $delete_newsletter_subscriber_template->delete();
+        if (!empty($delete_newsletter_subscriber_template)) {
+            $delete_newsletter_subscriber_template->delete();
+        }
         $newsletterTemplate->delete();
         return redirect()->route('newsletter-templates.index')->with('success', 'Newsletter template deleted successfully!');
     }
